starting to publicize

+++ __init__.py
@@ -6,9 +6,11 @@ from aqt.utils import showInfo
 import os
 import sys
 import asyncio
-sys.path.append(os.path.join(os.getenv('APPDATA'), 'Anki2', 'addons21', 'viet_dict'))
+
+print(os.getcwd())
+sys.path.append(os.getcwd())
 # Add the path to the src directory to sys.path
-sys.path.append(os.path.join(os.getenv('APPDATA'), 'Anki2', 'addons21', 'viet_dict', 'libs'))
+sys.path.append(os.path.join(os.getcwd(), 'libs'))
 
 # Now import load_dotenv from the dotenv package
 import dotenv
@@ -17,7 +19,6 @@ dotenv.load_dotenv()
 from translate import translate_text, tts
 
 note_type_name = "Sample"
-print(os.path.abspath(os.getcwd()))
 
 
 def start() -> None:
@@ -78,43 +79,50 @@ def on_add_cards_init(addcard):
             viet = None
             if("Viet" in get_field_values()):
                 viet = get_field_values()['Viet']
-                showInfo(viet)
             if viet:
-                print(f"{viet} in if")
                 loop = asyncio.get_event_loop()
                 async def fill_fields(viet, addcard):
-                    
-                    
-                    print("fill fields")
-                    print(viet)
                     head, _ = os.path.split(os.getcwd())
                     parent_dir, _ = os.path.split(head)
                     media = os.path.join(parent_dir, os.getenv("USER"), "collection.media")
                     viet = viet.strip()
 
                     note = addcard.editor.note
-                    english = await translate_text(viet)
-                    english = english["data"]["translations"]
-                    english = [translation["translatedText"] for translation in english]
-                    note.fields[1] = '\n'.join(english)
+                    try:
+                        english = await translate_text(viet)
+                        english = english["data"]["translations"]
+                        english = [translation["translatedText"] for translation in english]
+                        note.fields[1] = '\n'.join(english)
+                    except Exception as e:
+                        showInfo("Something went wrong\nYou might have to change the .env file")
+                        print(e)
+                        return False
                     
                     audio_file = f"{viet}.mp3"
                     audio_path = os.path.join(media, audio_file)
                     local_audio_path = os.path.join(parent_dir, "addons21", "viet_dict", "audio", audio_file)
                     media_output_file = os.path.join(media, audio_file)
-                    print(local_audio_path, media_output_file)
-                    try:
-                        os.rename(local_audio_path, media_output_file)
-                    except FileExistsError:
-                        print("File already exists")
-                    except Exception as e:
-                        print(e)
+
+                    if(not os.path.exists(media_output_file)):
+                        try:
+                            if(not os.path.exists(local_audio_path)):
+                                print("Going to Google TTS API")
+                                await tts(viet)
+                        except Exception as e:
+                            print(e)
+                            showInfo("Something went wrong\nYou might have to change the .env file")
+                            return False
+                        try:
+                            os.rename(local_audio_path, media_output_file)
+                            if(os.path.exists(local_audio_path)):
+                                print("removing")
+                                os.remove(local_audio_path)
+                        except FileExistsError:
+                            print("File already exists")
+                        except Exception as e:
+                            print("os.rename error")
+                            print(e)
                     note.fields[3] = f'[sound:{audio_file}]'
-                    try:
-                        await tts(viet)
-                    except Exception as e:
-                        print(e)
-                        return
                 
                 def success(addcard):
                     addcard.editor.loadNote()
